fix pedestrian simulator for ROS1 and ROS2 compatibility.

+++ include/pedestrian_simulator/configuration.h
@@ -52,7 +52,7 @@ public:
   Eigen::Matrix2d origin_R_;
 
   double prediction_step_;
-  double horizon_N_;
+  int horizon_N_;
 
   std::string scenario_file_;
